Remove lower-resolution bandwidth data that adds no information.
From GitHub issue #2: "We include bandwidth data for all time periods as
soon as we have at least two data points.  But sometimes a time period is
completely contained in another, shorter time period, where graphs have a
higher resolution.  For example, if a relay is only running for a few
months, we don't need to include bandwidth data for the 5-years graph, but
only for the 1-year graph.  In general, we only need to include bandwidth
data up to the point where the start of a time period precedes the first
bandwidth data point.  No need to include larger time periods."