Accept hashed fingerprints in search and lookup URLs.
More precisely, accept hashed relay fingerprints and hashed hashed bridge
fingerprints.

The problem addressed here is that both */lookup/ and */search/ URLs
require a non-hashed fingerprint to look up a relay and a hashed
fingerprint to look up a bridge.  This works fine if a) users know how to
hash fingerprints and b) the client knows whether the user provided a
hashed or non-hashed fingerprint.  But what if either a) or b) is not the
case?  If the client sees a 40-character fingerprint it shouldn't simply
submit it to the server.  It might be a non-hashed bridge fingerprint.

The change made here is to make */lookup/ and */search/ URLs accept hashed
relay fingerprints and hashed hashed bridge fingerprints.  Clients can now
be advised to always hash any 40-character hex input they receive from
users and include that in their request to the server.  If the user
provides a non-hashed fingerprint of a bridge the client won't reveal the
non-hashed fingerprint in the URL.  If the user provides a hashed
fingerprint, looking up the hash of the hash still works and returns the
bridge the user was looking for.

Implements #5368.