Don't use server descriptors parse histories.
If we parse a relay or bridge server descriptor that we never saw before
and learn about the relay or bridge in a later consensus or status, we'll
never write the descriptor content anywhere.  The result would be details
files containing no descriptor parts until the relay or bridge publishes
the next descriptor.  Yes, this took a long time to track down.