Add parameters for filtering and sorting results.
Parameters are more flexible than our current URLs.  Current functionality
can be implemented with the type, running, search, and lookup parameters.
The order and limit parameters can be used, e.g., for top-10 lists of
relays by bandwidth.  The offset parameter might be useful for paged
results.

This is a major API change, so we're leaving the deprecated methods in for
two months instead of one.