Store status bandwidth/weights files in subdirectories.
Right now, there are 300k bandwidth history files and 140k weights history
files in Onionoo's status directory.  weasel suggests to store those files
in subdirectories, e.g., 1/2/1234567.. to speed up access.  Let's do that.

The real fix is to use a database, of course.  But until we have that,
let's try to make the file system based solution not suck.