Fix off-by-one error in offset parameter.
With 10 relays and 10 bridges, setting an offset of 11 should have
returned 9 bridges (relays are removed first), but it return 10.  An
offset of 12 would have returned 9 bridges, and so on.

Here's why: we erroneously added a null value to the set of relays before
applying the offset parameter.  After throwing out the last relay we threw
out that null value instead of the first bridge.