Fix searching by hashed fingerprint.
We support searches for relays by hashed fingerprint and for bridges by
hashed hashed fingerprint.  The reason is that applications should always
hash full fingerprints in order not to accidentally leak non-hashed bridge
fingerprints.

However, the spec is vague about searching for beginnings of hashed relay
fingerprints and hashed hashed bridge fingerprints.  The current code did
not support those, but it should.  This commit changes that.