Make sure primary OR address is first in array.
Right now, relays and bridges have exactly one IPv4 address and at most
one IPv6 address.  Due to alphanumerical ordering the primary OR address
is first in the OR address array anyway.  But the requirements to IPv4 or
IPv6 addresses might change in the future, so make sure the ordering stays
this way.