Fix "running" fields in summary and details documents.
We say that a bridge is "running" if it's contained in the bridge network
status, but it also needs the "Running" flag.

This worked fine for relays, because only running relays are contained in
the consensus, but bridge network statuses may contain non-running
bridges.

We should require the "Running" flag for both relays and bridges, just in
case we fall back to a consensus method that contains non-running relays.

Fixes #9198.