Merge two writer classes to speed up rDNS lookups.
The NodeDataWriter and DetailsDataWriter did very similar things.  They
processed various descriptors and performed GeoIP and rDNS lookups to
learn the latest status of a relay or bridge for summary and details
documents.  Merging these classes makes sense, because we can start rDNS
lookups much earlier which then run in the background.