Stop deleting details, bandwidth, and weights documents.
So far, we deleted details, bandwidth, and weights documents when a relay
or bridge went offline for at least 7 days.  The reason was that we
wouldn't return documents for these relays or bridges anyway, and once the
relay or bridge comes back online, we'd write a new document from our
status files.

Very soon we'll want to support searching for relays and bridges that have
been offline for more than 7 days, and we'll also want to serve their
details, bandwidth, and weights documents.  We need to stop deleting them.

Related to this, we now write new details, bandwidth, and weights
documents for all relays and bridges, regardless of whether they were
running in the last 7 days.

This step also removes some dependencies in the code which may allow us to
parallelize more stuff in the future.