Be more precise about leaving out non-null values.
The specification said that we're only including non-null values for
series of at least two subsequent data points.  But that's not true.
We're not including histories if there are not at least two subsequent
non-null values.  But those histories may contain single non-null values
with null values previous and next to them.