Fix compressing bandwidth and weights histories.
We compress histories by merging adjacent intervals to save disk space,
unless we want intervals to stay distinct.  For example, we might want to
compress all intervals on the same UTC day but not want to merge intervals
with the previous or next UTC day.  However, we had an off-by-one error
which made us merge the wrong intervals.  For example, we merged intervals
from 23:00:00 on one day to 23:00:00 the next day.