Don't share SimpleDateFormat instance which is not thread-safe.
When we moved date parsing and formatting to a util class in 14a6e59, we
didn't think of thread-safety of the SimpleDateFormat class.  Turns out
it's *not* thread-safe.

The hot fix is to create a new SimpleDateFormat instance for each request.

A better solution might be to share a single instance per date format and
thread, or to share an instance of a thread-safe SimpleDateFormat
alternative per date format.  Added these as TODOs.

Fixes #11516.