Add a new fingerprint parameter.
By default, we are limiting all responses to relays or bridges that have
been running in the past 7 days.  The main reason for this is performance.

We're now making an exception for specific lookups by fingerprint or
hashed fingerprint.  These lookups don't use the node index, but attempt
to read the relay's or bridge's details document and extract all
information needed to compile a summary document or locate one of the
other documents.

There may be more elegant solutions, but we'll probably have to switch to
a database design for them.

Implements #11350.