Fix contact parameter.
We support searching for relays by contact.  In order to learn contacts,
we need to parse contact lines from server descriptors.  However, our
earlier approach, which was introduced in 51d883b, only stored the contact
whenever the relay published a new server descriptor.  In all other
executions we would reset that contact line, assuming that the relay did
not provide a contact line.

The new approach is to rely on the details status file that we're writing
for details documents and use the contact information contained there.
This also works better than the earlier approach, because the details
status file also contains the descriptor publication time which allows us
to always memorize the latest contact information of a relay.

Fixes #11822.