Avoid storing full consensuses in memory.
We stored full consensuses in memory, so that we could compute weights
when all descriptors have been parsed.  Obviously, this does not scale
with more than a few days of consensuses.

The new approach is to parse server descriptors *before* consensuses and
store advertised bandwidths in weights status files when doing so.  Later,
when we parse consensuses, we are sure that if there's a server descriptor
for a relay, it'll be contained in the weights status file.