Download input descriptors from CollecTor via https.
We used to rsync descriptors, but that was only a good idea when source
and destination were on the same host.  As weasel points out, rsync
provides no server authentication, so we don't know what we get.

Another advantage is that all necessary steps for the hourly execution are
now implemented in Java.  We might now consider to move to a different
execution model than an hourly cronjob much more easily.