Don't compress weights values with missing values.
When compressing weights status lines, we simply combined adjacent
intervals by calculating arithmetic means.  But we did not consider
missing values, which were encoded as 0.0.  As a result the result was
lower than it should really have been.  Now that missing values are
encoded as -1.0, it becomes even more apparent that this approach was
wrong with negative weights values of, e.g., -0.127320915201.  As a fix,
avoid compressing intervals with missing values for certain weights.

Example of an old combined line:

2008-08-20 20:00:00 2008-08-29 00:00:00 0.000261225507 -0.127320915201
  -1.000000000000 -1.000000000000 -1.000000000000 76374.857142857120
  65.306122448979

The -0.127320915201 part is wrong there.

Example of the same data, not combined:

2008-08-20 20:00:00 2008-08-21 21:00:00 0.000245451671 -1.000000000000
  -1.000000000000 -1.000000000000 -1.000000000000 73466.880000000000
  -1.000000000000
2008-08-21 21:00:00 2008-08-29 00:00:00 0.000263531623 0.000263746320
  -1.000000000000 -1.000000000000 -1.000000000000 76800.000000000000
  75.000000000000

Found in the context of implementing #11388.