Keep at most one parsed descriptor in the queue.
When we switched to parsing concatenated server and extra-info descriptors
we limited the descriptor queue to at most one parsed descriptor file.
Actually, this also makes sense for large descriptors like consensuses.
There's no need to keep up to 100 of them in the queue, because it takes
much longer to process them than to parse them.

A quick look at the logs confirms that, with setting the max queue length
to 1, we almost never have to wait for a descriptor provided by
metrics-lib.