Move front-end parts of NodeStatus to SummaryDocument.
The NodeStatus class was seriously overloaded:

 1. NodeDetailsStatusUpdater uses the list of NodeStatus instances as
    input to its GeoIP and rDNS look-up operations and to calculate
    path-selection probabilities.

 2. DetailsDocumentWriter uses fields from NodeStatus instances (some of
    them being stored persistently, some only kept in memory) to write
    part of DetailsDocument instances.

 3. NodeIndexer obtains all relevant data to populate the search index
    from NodeStatus instances.

 4. ResponseBuilder uses NodeStatus instances to produce summary documents
    on-the-fly.

As first step to reduce this complexity, let's copy the parts used for 3
and 4 to a currently unused class, SummaryDocument, and use that class in
NodeIndexer and ResponseBuilder.  It makes sense to combine those two
purposes for performance reasons.  Parts 1 and 2 will be simplified in
later commits.