Add explicit way to invalidate document cache.
There are two document types that are fully cached in memory for
performance reasons: NodeStatus and SummaryDocument.

Before 6854b64 we initialized (or re-initialized) this cache when the user
called the list() operation.  That didn't seem very robust, because it was
an implicit protocol the user had to follow: if they didn't call list(),
but retrieve(), they would have run into an exception.  And if they called
list(), changed documents in the cache, did not call flushDocumentCache()
but instead called list() again, all their changes would have been lost.

That's why we changed all methods in DocumentStore to initialize the cache
in case that hasn't been done before.

However, and here's the bug, subsequent calls to list() wouldn't update
the cache anymore, which the servlet implicitly relied on.

The fix is to add an explicit method invalidateDocumentCache() that clears
all local caches.  The next call to list() or any other method will
re-initialize the cache.

Fixes #12797.