Add new method to list recently stored documents.
This is a rather big refactoring of how document writers obtain a list of
documents they should write.

The old way was to register a "fingerprint listener" at the descriptor
source and write documents of all relays and bridges whose fingerprints
it they learned about.  One drawback (or even bug; cf. #12651) was that
document writers didn't learn about relays or bridges that went offline,
because there's no descriptor for that.

The new way is for document writers to ask for a list of status files that
have changed since the last "update status" file was written.  This can be
implemented for documents cached in memory and for documents stored on
disk.  Document writers can then use this list to only write the
documents that need updating.