Don't escape `/` before feeding strings into Gson.
We escape a few characters in strings before feeding them into Gson,
including backslashes, double quotes, \b, \n, \t, \f, \r, and non-ASCII
characters.  But we also escape `/` to `\/` which we did not change back
after receiving Gson's output.  We should simply stop escaping forward
slashes, because Gson leaves them untouched.  The easiest way to change
this is to switch from escapeEcmaScript to escapeJava which also saves us
the effort of undoing the escaping of `'`.

Fixes #13267.