Avoid error because of non-existant descriptor type
When splitting up the hourly updater into three phases, we lost the
ability to skip downloading descriptors we're not going to process
later.  Hence, we'll get an error because CollecTor doesn't serve
bridge-pool assignments anymore.  But we don't parse them anyway, so
we can as well remove them from our list of descriptors we might care
about.  That avoids this specific error.