Remove deprecated "family" field from details documents.
On August 25, 2015, we added the optional "alleged_family" and
"effective_family" fields and deprecated the likewise optional
"family" field in details documents.  As of yesterday, Atlas (#16961),
Globe (#16962), and Compass (#17720) stopped using the "family" field
and switched over to the new fields.  It's time to finally remove the
"family" field from details documents.

Implements #18026.