Stop returning code 500 when serving stale data.
Until now, when the hourly updater has not run for 6 hours or more,
the server responds to all queries with a 500 Internal Server Error.
This seemed to be a good idea, because users would tell us when the
hourly updater was broken and we could fix it.  But actually, these 6
hours were chosen arbitrarily, and each application should decide for
itself when it considers data as stale.  That's why we're taking out
the check and stopping to return a 500 status code because of
potentially stale data.

Implements #16907.