Fix invalid first_seen timestamps in existing data.
In a previous commit we fixed an issue with the first_seen timestamp
of bridges when parsing new bridge descriptors.  We still need to fix
invalid first_seen timestamps in existing data.

There are two different cases of invalid first_seen timestamps in
existing data:

1) first_seen is 1970-01-01 00:00:00, last_seen is 2014-12-09 or
later.  In these cases we parsed a server descriptor before first
seeing the bridge in a status in a subsequent execution.  We need to
special-case the previously known first_seen timestamp 1970-01-01
00:00:00 (value 0) and not consider it earlier than an actual
timestamp.

2) Both first_seen and last_seen are 1970-01-01 00:00:00.  We have
never seen these bridges in a status.  We can safely discard these
entries while reading them from disk to memory.

With this fix it will be possible to repair the status data of an
existing Onionoo instance by re-importing archived bridge network
statuses from 2014-12 and onwards.

Fixes the remainder of #20994.