Fix the "running" field in bridge details documents.
When we set this field in the update process, we only looked whether
the bridge is contained in the last known bridge network status. We
also need to check whether it has the "Running" flag assigned there.

This is different from relays, because the consensus only lists relays
with the "Running" flag since a couple of years, whereas the bridge
network status lists all known bridges.

Fixes #23467. Spotted by nusenu.