Increase test timeout from 100 to 200ms.
The testCountryDeDe typically requires around 80 to 100ms to complete,
which makes our test timeout of 100ms somewhat optimistic. The reason
is that it's the first test method that is run in this test class, and
it takes some time to initialize the test environment. This is just a
hotfix, but it's also not really a serious problem, so maybe it'll
just do.