Schedule next major version.
Optimistically schedule the next major version for end of this month
for removing $ from family fingerprints. That's a rather small
backward-incompatible change, so it should be fine to reduce the
notification period to a bit under four weeks. But just in case we
want more time we can always push that back by a week or two. At least
we're now saying that there will be a major version update soon.