Don't rely on system time for compressing histories.
Related to not relying on system time for writing histories, we also
shouldn't rely on it for compressing histories anymore. Otherwise we
might be compressing the history of a relay or bridge that recently
went offline while expecting an uncompressed history of that relay for
writing history documents. In short, let's do the same when updating
*Status objects as we do when writing *Document objects.

Related to #16513, but not strictly part of it.