Use the same last-seen time for writing documents.
Previously, we'd have used the last-seen time of the relay/bridge that
we're writing a document for. But that doesn't make much sense,
because we don't have to provide, e.g., a 1 week graph for a relay
that has been offline for years. This change is going to shrink the
out/ directory to a similar size as it was when we used current system
time for writing documents.