Change 3 month weights graph to 24 hours detail.
We recently changed the 3 month bandwidth graph to 24 hours detail,
and we changed the compression of various status files to support 6
month graphs with 24 hours detail. However, we did not consider that
we currently need weights data for the last 3 months on a detail of 12
hours. This commit adapts the output file by using the data that we
now have. Maybe it even makes sense to use the same detail in both
graphs, though that was not the original intention.