Stop sharing the JNDI DirContext
In the implementation of more accurate DNS resolution, the DirContext
used by JNDI lookups was shared in a utility class. It turns out that
InitialDirContext, the implementation we used, is very not thread-safe
and so as a result lookups would often fail to complete.

This commit creates a new DirContext for every lookup. The overhead in
doing so is minimal and ensures that we do not have thread safety
issues.

This commit also increases RDNS_LOOKUP_WORKERS_NUM to 30 (from 5). The
majority of time in these threads is spent waiting for replies and so
there are gains to be had in increasing the number of requests in
flight.

Fixes: #26963